fixed some crashes, added calibration and lolin32 (limited testing)

+++ src/headers/blynk.h
@@ -1,9 +1,7 @@
 #ifndef BLYNK_H
     #define BLYNK_H
     #include <Arduino.h>  // for type definitions
-    
     void blynkSetup();
-    void blynkLoop(int sensorReading);
+    void blynkLoop(float sensorReading);
     void println(String message);
-
 #endif
